+++ src/NServiceBus.Core/Routing/FileBasedRoundRobinRoutingDistributor.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.Unicast.Routing
+namespace NServiceBus.Routing
 {
     using System;
     using System.Collections.Concurrent;
