+++ src/NServiceBus.Core/Routing/RoutingDistributorDefinition.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.Unicast.Routing
+namespace NServiceBus.Routing
 {
     using System;
